clean up code. remove talon config from the command init function in the liftMotor subsystem. Put them in the constructor for now. Use Phoenix tuner for on the fly configuration from now on.

+++ 2019-master/src/main/java/org/usfirst/frc/team4970/robot/commands/LiftToBallIntakePosition.java
@@ -36,7 +36,7 @@ public class LiftToBallIntakePosition extends Command {
     protected boolean isFinished() {
     	if ((isTimedOut()) || 
     		((Math.abs(Robot._liftMotor.getEncoderCount() - Constants.liftBallIntakePositionSetpoint))
-    			<= (int)Constants.liftMotorAllowableClosedLoopError))
+    			<= Constants.liftMotorAllowableClosedLoopError))
     	{
     		LiftMotor._liftState = LiftMotor.LiftState.LIFT_INTAKE_HEIGHT;
     		return true;
