+++ 2019-master/src/main/java/org/usfirst/frc/team4970/robot/commands/LiftToHatchLowerPosition.java
@@ -36,7 +36,7 @@ public class LiftToHatchLowerPosition extends Command {
     protected boolean isFinished() {
     	if ((isTimedOut()) || 
     		((Math.abs(Robot._liftMotor.getEncoderCount() - Constants.liftHatchPlaceLowerPositionSetpoint))
-    			<= (int)Constants.liftMotorAllowableClosedLoopError))
+    			<= Constants.liftMotorAllowableClosedLoopError))
     	{
     		LiftMotor._liftState = LiftMotor.LiftState.LIFT_HATCH_HEIGHT;
     		return true;
